+++ example/lib/main.dart
@@ -1,4 +1,6 @@
+import 'dart:io';
 import 'package:flutter/material.dart';
+import 'package:flutter_upgrader/flutter_upgrader.dart';
 
 void main() {
   runApp(const MyApp());
@@ -19,7 +21,50 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Flutter Upgrader Plugin'),
         ),
-        body: Container(),
+        body: Padding(
+          padding: const EdgeInsets.only(top: 28.0),
+          child: Row(
+            children: [
+              Expanded(
+                child: ElevatedButton(
+                  style: ElevatedButtonTheme.of(context).style,
+                  child: const Padding(
+                    padding: EdgeInsets.all(12.0),
+                    child: Text(
+                      '升 级',
+                      style: TextStyle(fontSize: 18),
+                    ),
+                  ),
+                  onPressed: () {
+                    final Future<AppUpgradeInfo> appUpgradeInfo = Future.value(
+                      AppUpgradeInfo(
+                        title: '更新提示',
+                        contents: ['有新版本哟,请更新～'],
+                        force: true,
+                      ),
+                    );
+
+                    if (Platform.isAndroid) {
+                      AppUpgradeManager.upgrade(
+                        context,
+                        appUpgradeInfo,
+                        appMarketInfo: AppMarketManager.huaWei,
+                      );
+                    }
+
+                    if (Platform.isIOS) {
+                      AppUpgradeManager.upgrade(
+                        context,
+                        appUpgradeInfo,
+                        iosAppId: 'idxxxxxx',
+                      );
+                    }
+                  },
+                ),
+              ),
+            ],
+          ),
+        ),
       ),
     );
   }